requirements.txt
arrow==0.15.5
bleach==3.1.0
certifi==2019.11.28
chardet==3.0.4
idna==2.8
Jinja2==2.10.3
Markdown==3.1.1
MarkupSafe==1.1.1
mdx-linkify==1.2
Pillow==7.0.0
pkg-resources==0.0.0
python-dateutil==2.8.1
python-slugify==4.0.0
requests==2.22.0
six==1.13.0
text-unidecode==1.3
urllib3==1.25.7
webencodings==0.5.1
xmltodict==0.12.0
